Css added to site

+++ src/Components/Command.tsx
@@ -11,24 +11,28 @@ const Command:React.FC<Props> = ({selectedItem}) => {
     
     
     const deleteCommand = () => {
-        fetch("http://192.168.1.153:5555/deletepost/"+selectedItem.id,{
-            method: "DELETE"
-        })
-        .then(res => { return res.json() })
-        .then(data => console.log(data))
-        .catch(err => console.log("Error", err))
+        if(window.confirm('DELETE POST ?')){
+            fetch("http://192.168.1.153:5555/deletepost/"+selectedItem.id,{
+                method: "DELETE"
+            })
+            .then(res => { return res.json() })
+            .then(data => console.log(data))
+            .catch(err => console.log("Error", err))
+        }
     }
 
     if(allData)
     return(
-        <div>
-            <h2>Name: {allData.name}</h2>
-            <h3>Platform: {allData.platform}</h3>
-            <p>Description: {allData.description}</p>
-            <p>Can be used with: {allData.options}</p>
-            <p>Examples: {allData.howTo}</p>
-            <button>Edit command</button>
-            <button onClick={() => deleteCommand()} >DELETE COMMAND</button>
+        <div className="command_background">
+            <p className = "command_topic">{allData.name}</p>
+            <p className = "command_platform">{allData.platform}</p>
+            <p className = "command_description" >{allData.description}</p>
+            <p className = "command_options" >{allData.options}</p>
+            <p className = "command_example" >{allData.howTo}</p>
+            <div className="command_buttons_div">
+                <button className="button" >Edit</button>
+                <button className="button" onClick={() => deleteCommand()} >DELETE</button>
+            </div>
         </div>
     )
